+++ JotunnLib/Main.cs
@@ -21,7 +21,7 @@ namespace Jotunn
         /// <summary>
         ///     The current version of the Jotunn library.
         /// </summary>
-        public const string Version = "2.11.6";
+        public const string Version = "2.11.7";
 
         /// <summary>
         ///     The name of the library.