+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Changed empty CustomLocalization constructor to be marked obsolete, LocalizationManager.Instance.GetLocalization() should be used instead
 * Slightly improved loading time of big mod packs when creating custom entities again
 
+## Version 2.11.7
+* Added Valheim network version check to the disconnect window to better identify the cause of a disconnect. Only visible if both server and client are running Jotunn 2.11.7 or higher
+* Removed the Valheim version string check from the disconnect window, as the network version is used for the version check
+
 ## Version 2.11.6
 * Fixed an issue with custom category tabs where the hammer menu wouldn't open when using a controller